Remove pre-seeding prime number generator
Remove the code that supported pre-seeding the prime number generator
with already generated primes.  It wasn't a bad idea, but the
implementation didn't work (it failed to increment the composite
generators to a sensible starting point if they were starting at a value
lower than the biggest known prime).  Possibly something to add back
later, but for now it's unused so just remove it.